Move around comp tests and tweak coverage
