+++ platform-operator/controllers/verrazzano/component/certmanager/certmanager_test.go
@@ -5,9 +5,7 @@ package certmanager
 
 import (
 	"context"
-	"github.com/verrazzano/verrazzano/pkg/log/vzlog"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
 	corev1 "k8s.io/client-go/kubernetes/typed/core/v1"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"testing"
@@ -361,189 +359,6 @@ func certificateExists(client clipkg.Client, name string, namespace string) (boo
 		return false, clipkg.IgnoreNotFound(err)
 	}
 	return true, nil
-}
-
-// TestPostInstallCA tests the PostInstall function
-// GIVEN a call to PostInstall
-//  WHEN the cert type is CA
-//  THEN no error is returned
-func TestPostInstallCA(t *testing.T) {
-	localvz := vz.DeepCopy()
-	localvz.Spec.Components.CertManager.Certificate.CA = ca
-
-	defer func() { getClientFunc = GetCoreV1Client }()
-	getClientFunc = createClientFunc(localvz.Spec.Components.CertManager.Certificate.CA)
-
-	client := fake.NewFakeClientWithScheme(testScheme)
-	err := fakeComponent.PostInstall(spi.NewFakeContext(client, localvz, false))
-	assert.NoError(t, err)
-}
-
-// TestPostUpgradeUpdateCA tests the PostUpgrade function
-// GIVEN a call to PostUpgrade
-//  WHEN the type is CA and the CA is updated
-//  THEN the ClusterIssuer is updated correctly and no error is returned
-func TestPostUpgradeUpdateCA(t *testing.T) {
-	runCAUpdateTest(t, true)
-}
-
-// TestPostInstallUpdateCA tests the PostInstall function
-// GIVEN a call to PostInstall
-//  WHEN the type is CA and the CA is updated
-//  THEN the ClusterIssuer is updated correctly and no error is returned
-func TestPostInstallUpdateCA(t *testing.T) {
-	runCAUpdateTest(t, false)
-}
-
-func runCAUpdateTest(t *testing.T, upgrade bool) {
-	localvz := vz.DeepCopy()
-	localvz.Spec.Components.CertManager.Certificate.CA = ca
-
-	updatedVZ := vz.DeepCopy()
-	newCA := vzapi.CA{
-		SecretName:               "newsecret",
-		ClusterResourceNamespace: "newnamespace",
-	}
-	updatedVZ.Spec.Components.CertManager.Certificate.CA = newCA
-
-	defer func() { getClientFunc = GetCoreV1Client }()
-	getClientFunc = createClientFunc(updatedVZ.Spec.Components.CertManager.Certificate.CA)
-
-	expectedIssuer := &certv1.ClusterIssuer{
-		Spec: certv1.IssuerSpec{
-			IssuerConfig: certv1.IssuerConfig{
-				CA: &certv1.CAIssuer{
-					SecretName: newCA.SecretName,
-				},
-			},
-		},
-	}
-
-	client := fake.NewFakeClientWithScheme(testScheme, localvz)
-	ctx := spi.NewFakeContext(client, updatedVZ, false)
-
-	var err error
-	if upgrade {
-		err = fakeComponent.PostInstall(ctx)
-	} else {
-		err = fakeComponent.PostInstall(ctx)
-	}
-	assert.NoError(t, err)
-
-	actualIssuer := &certv1.ClusterIssuer{}
-	assert.NoError(t, client.Get(context.TODO(), types.NamespacedName{Name: verrazzanoClusterIssuerName}, actualIssuer))
-	assert.Equal(t, expectedIssuer.Spec.CA, actualIssuer.Spec.CA)
-}
-
-// TestPostInstallAcme tests the PostInstall function
-// GIVEN a call to PostInstall
-//  WHEN the cert type is Acme
-//  THEN no error is returned
-func TestPostInstallAcme(t *testing.T) {
-	localvz := vz.DeepCopy()
-	localvz.Spec.Components.CertManager.Certificate.Acme = acme
-	client := fake.NewFakeClientWithScheme(testScheme)
-	// set OCI DNS secret value and create secret
-	localvz.Spec.Components.DNS = &vzapi.DNSComponent{
-		OCI: &vzapi.OCI{
-			OCIConfigSecret: "ociDNSSecret",
-			DNSZoneName:     "example.dns.io",
-		},
-	}
-	client.Create(context.TODO(), &v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "ociDNSSecret",
-			Namespace: ComponentNamespace,
-		},
-	})
-	err := fakeComponent.PostInstall(spi.NewFakeContext(client, localvz, false))
-	assert.NoError(t, err)
-}
-
-// TestPostUpgradeAcmeUpdate tests the PostUpgrade function
-// GIVEN a call to PostUpgrade
-//  WHEN the cert type is Acme and the config has been updated
-//  THEN the ClusterIssuer is updated as expected no error is returned
-func TestPostUpgradeAcmeUpdate(t *testing.T) {
-	runAcmeUpdateTest(t, true)
-}
-
-// TestPostInstallAcme tests the PostInstall function
-// GIVEN a call to PostInstall
-//  WHEN the cert type is Acme and the config has been updated
-//  THEN the ClusterIssuer is updated as expected no error is returned
-func TestPostInstallAcmeUpdate(t *testing.T) {
-	runAcmeUpdateTest(t, false)
-}
-
-func runAcmeUpdateTest(t *testing.T, upgrade bool) {
-	localvz := vz.DeepCopy()
-	localvz.Spec.Components.CertManager.Certificate.Acme = acme
-	// set OCI DNS secret value and create secret
-	oci := &vzapi.OCI{
-		OCIConfigSecret: "ociDNSSecret",
-		DNSZoneName:     "example.dns.io",
-	}
-	localvz.Spec.Components.DNS = &vzapi.DNSComponent{
-		OCI: oci,
-	}
-
-	oldSecret := &v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "ociDNSSecret",
-			Namespace: ComponentNamespace,
-		},
-	}
-
-	newSecret := &v1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "newociDNSSecret",
-			Namespace: ComponentNamespace,
-		},
-	}
-
-	existingIssuer, _ := createAcmeClusterIssuer(vzlog.DefaultLogger(), templateData{
-		Email:       acme.EmailAddress,
-		Server:      acme.Environment,
-		SecretName:  oci.OCIConfigSecret,
-		OCIZoneName: oci.DNSZoneName,
-	})
-
-	updatedVz := vz.DeepCopy()
-	newAcme := vzapi.Acme{
-		Provider:     "letsEncrypt",
-		EmailAddress: "slbronkowitz@gmail.com",
-		Environment:  "production",
-	}
-	newOCI := &vzapi.OCI{
-		DNSZoneCompartmentOCID: "somenewocid",
-		OCIConfigSecret:        newSecret.Name,
-		DNSZoneName:            "newzone.dns.io",
-	}
-	updatedVz.Spec.Components.CertManager.Certificate.Acme = newAcme
-	updatedVz.Spec.Components.DNS = &vzapi.DNSComponent{OCI: newOCI}
-
-	expectedIssuer, _ := createAcmeClusterIssuer(vzlog.DefaultLogger(), templateData{
-		Email:       newAcme.EmailAddress,
-		Server:      letsEncryptProdEndpoint,
-		SecretName:  newOCI.OCIConfigSecret,
-		OCIZoneName: newOCI.DNSZoneName,
-	})
-
-	client := fake.NewFakeClientWithScheme(testScheme, localvz, oldSecret, newSecret, existingIssuer)
-	ctx := spi.NewFakeContext(client, updatedVz, false)
-
-	var err error
-	if upgrade {
-		err = fakeComponent.PostInstall(ctx)
-	} else {
-		err = fakeComponent.PostInstall(ctx)
-	}
-	assert.NoError(t, err)
-
-	actualIssuer, _ := createAcmeClusterIssuer(vzlog.DefaultLogger(), templateData{})
-	assert.NoError(t, client.Get(context.TODO(), types.NamespacedName{Name: verrazzanoClusterIssuerName}, actualIssuer))
-	assert.Equal(t, expectedIssuer.Object["spec"], actualIssuer.Object["spec"])
 }
 
 // fakeBash verifies that the correct parameter values are passed to upgrade
